refactor(c/validation): split up large test file
Fixes #1237.
